// APP/src/app/common/prod-lines.enum.ts
export enum ProductionLines {
  ProHS = 'HIPPO SAK EXTRUSION',
  ProMR = 'HIPPO SAK EXTRUSION',
  ProStripCut = 'HIPPO SAK STRIP',
  ProMasterStrip = 'HIPPO SAK STRIP',
  ProPNP = 'PULL-N-PAK',
  ProRepro = 'REPRO',
  ProTB = 'TRASH BAG CONV',
  ProPalletization = "PROPALLETIZATION",
  ProScrap = "SCRAP",
  ProSuper = "PROSUPER"
}